:white_check_mark: batch 11: edition pairing tier 1, remote workers + transcode, challenge_solve browser worker, archive-drop ingest + audio items, ABS official app round 2, editions design; registry installed in routers::mount

+++ crates/abs/src/routes/session.rs
@@ -446,7 +446,7 @@ pub(crate) fn stats(sessions: Vec<PlaybackSessionDto>) -> ListeningStatsDto {
 	let mut total = 0.0;
 
 	for session in &sessions {
-		let listened = session.time_listening;
+		let listened = session.time_listening as f64;
 		total += listened;
 		*days.entry(session.date.clone()).or_default() += listened;
 		*day_of_week.entry(session.day_of_week.clone()).or_default() += listened;
